requirements.txt
anyio==3.6.1
asgiref==3.5.2
blis==0.7.7
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
decorator==5.1.1
fastapi==0.78.0
filelock==3.6.0
gensim==4.2.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
h11==0.13.0
huggingface-hub==0.5.1
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.22.4
packaging==21.3
pathy==0.6.1
Pillow==9.1.0
preshed==3.0.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.9
python-multipart==0.0.5
PyYAML==6.0
regex==2022.4.24
requests==2.26.0
rsa==4.8
sacremoses==0.0.53
scipy==1.8.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
starlette==0.19.1
thinc==8.0.17
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.18.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
validators==0.18.2
wasabi==0.9.1
weaviate-client==3.2.2
